Add EMR job spawning (and waiting) support.

+++ src/main/scala/org/broadinstitute/dig/aggregator/core/AWS.scala
@@ -1,9 +1,8 @@
 package org.broadinstitute.dig.aggregator.core
 
-import scala.collection.JavaConverters._
-
 import cats._
 import cats.effect._
+import cats.effect.concurrent._
 import cats.implicits._
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider
@@ -14,13 +13,18 @@ import com.amazonaws.services.s3.model._
 import com.amazonaws.services.elasticmapreduce._
 import com.amazonaws.services.elasticmapreduce.model.AddJobFlowStepsRequest
 import com.amazonaws.services.elasticmapreduce.model.AddJobFlowStepsResult
-import com.amazonaws.services.elasticmapreduce.model.HadoopJarStepConfig
-import com.amazonaws.services.elasticmapreduce.model.StepConfig
+import com.amazonaws.services.elasticmapreduce.model.JobFlowDetail
+import com.amazonaws.services.elasticmapreduce.model.ListStepsRequest
+import com.amazonaws.services.elasticmapreduce.model.StepState
+import com.amazonaws.services.elasticmapreduce.model.StepSummary
 
 import com.typesafe.scalalogging.LazyLogging
 
+import java.io.InputStream
+
+import scala.collection.JavaConverters._
+import scala.concurrent.duration._
 import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.Future
 
 /**
  * AWS controller (S3 + EMR clients).
@@ -31,19 +35,19 @@ final class AWS(config: BaseConfig) extends LazyLogging {
   /**
    * The same region and bucket are used for all operations.
    */
-  private val region: Regions = Regions.valueOf(config.aws.region)
-  private val bucket          = config.aws.s3.bucket
+  val region: Regions = Regions.valueOf(config.aws.region)
+  val bucket          = config.aws.s3.bucket
 
   /**
    * AWS IAM credentials provider.
    */
-  private val credentials: AWSStaticCredentialsProvider = new AWSStaticCredentialsProvider(
+  val credentials: AWSStaticCredentialsProvider = new AWSStaticCredentialsProvider(
     new BasicAWSCredentials(config.aws.key, config.aws.secret))
 
   /**
    * S3 client for storage.
    */
-  private val s3: AmazonS3 = AmazonS3ClientBuilder.standard
+  val s3: AmazonS3 = AmazonS3ClientBuilder.standard
     .withRegion(region)
     .withCredentials(credentials)
     .build
@@ -51,7 +55,7 @@ final class AWS(config: BaseConfig) extends LazyLogging {
   /**
    * EMR client for running map/reduce jobs.
    */
-  private val emr: AmazonElasticMapReduce = AmazonElasticMapReduceClientBuilder.standard
+  val emr: AmazonElasticMapReduce = AmazonElasticMapReduceClientBuilder.standard
     .withCredentials(credentials)
     .withRegion(region)
     .build
@@ -71,7 +75,14 @@ final class AWS(config: BaseConfig) extends LazyLogging {
   }
 
   /**
-   * Download a file from an S3 bucket.
+   * Upload a file to S3 in a particular bucket.
+   */
+  def put(key: String, stream: InputStream): IO[PutObjectResult] = IO {
+    s3.putObject(bucket, key, stream, new ObjectMetadata())
+  }
+
+  /**
+   * Fetch a file from an S3 bucket (does not download content).
    */
   def get(key: String): IO[S3Object] = IO {
     s3.getObject(bucket, key)
@@ -85,18 +96,29 @@ final class AWS(config: BaseConfig) extends LazyLogging {
   }
 
   /**
+   * List all the keys in a given S3 folder.
+   */
+  def ls(key: String): IO[List[String]] = IO {
+    s3.listKeys(bucket, key).keys
+  }
+
+  /**
    * Delete (recursively) all the keys under a given key from S3.
    */
   def rmdir(key: String): IO[List[String]] = {
     val ios = for (listing <- s3.listKeys(bucket, key)) yield {
-      val keys        = listing.getObjectSummaries.asScala.map(_.getKey)
-      val keyVersions = keys.map(new DeleteObjectsRequest.KeyVersion(_))
-      val request     = new DeleteObjectsRequest(bucket).withKeys(keyVersions.asJava)
+      if (listing.getObjectSummaries.size == 0) {
+        IO(Nil)
+      } else {
+        val keys        = listing.getObjectSummaries.asScala.map(_.getKey)
+        val keyVersions = keys.map(new DeleteObjectsRequest.KeyVersion(_))
+        val request     = new DeleteObjectsRequest(bucket).withKeys(keyVersions.asJava)
 
-      for {
-        _ <- IO(logger.debug(s"Deleting ${keys.head} + ${keys.tail.size} more keys..."))
-        _ <- IO(s3.deleteObjects(request))
-      } yield keys
+        for {
+          _ <- IO(logger.debug(s"Deleting ${keys.head} + ${keys.tail.size} more keys..."))
+          _ <- IO(s3.deleteObjects(request))
+        } yield keys
+      }
     }
 
     // all the delete operations can happen in parallel
@@ -116,26 +138,87 @@ final class AWS(config: BaseConfig) extends LazyLogging {
   }
 
   /**
-   * Run a map/reduce job.
+   * Given a set of steps, start a job.
    */
-  def runMR(jar: String,
-            mainClass: String,
-            args: Seq[String] = List.empty): IO[AddJobFlowStepsResult] = {
-    val stepConfig = new HadoopJarStepConfig()
-      .withJar(s"s3://${bucket}/jobs/$jar")
-      .withMainClass(mainClass)
-      .withArgs(args.asJava)
-
-    // create the step to run this config
-    val step = new StepConfig(mainClass, stepConfig)
-
-    // create the request to run the step
+  def runJob(steps: Seq[JobStep]): IO[AddJobFlowStepsResult] = {
     val request = new AddJobFlowStepsRequest()
       .withJobFlowId(config.aws.emr.cluster)
+      .withSteps(steps.map(_.config).asJava)
 
-    // start it
     IO {
-      emr.addJobFlowSteps(request)
+      val job = emr.addJobFlowSteps(request)
+
+      // show the job ID so it can be referenced in the AWS console
+      logger.debug(s"Submitted job with ${steps.size} steps...")
+      job
     }
+  }
+
+  /**
+   * Every 20 seconds, send a request to the cluster to determine the state
+   * of all the steps in the job. Only return once the state is one of the
+   * following for the last/current step:
+   *
+   *   StepState.COMPLETED
+   *   StepState.FAILED
+   *   StepState.INTERRUPTED
+   *   StepState.CANCELLED
+   */
+  def waitForJob(job: AddJobFlowStepsResult): IO[Either[StepSummary, Unit]] = {
+    import Implicits._
+
+    val request = new ListStepsRequest()
+      .withClusterId(config.aws.emr.cluster)
+      .withStepIds(job.getStepIds)
+
+    // wait a little bit then request status
+    val req = for (_ <- IO.sleep(20.seconds)) yield emr.listSteps(request)
+
+    /*
+     * The step summaries are returned in reverse order. The job is complete
+     * when all the steps have their status set to COMPLETED, at which point
+     * the curStep will be None.
+     */
+    val curStep = req.map(_.getSteps.asScala).map(_.find(!_.isComplete))
+
+    /*
+     * If the current step has failed, interrupted, or cancelled, then the
+     * entire job is also considered failed, and return the failed step.
+     */
+    curStep.flatMap {
+      case None => IO(Right(()))
+
+      // the current step stopped for some reason
+      case Some(step) if step.isStopped =>
+        logger.error(s"Job failed: ${step.stopReason}")
+        IO(Left(step))
+
+      // still waiting for the current step to complete
+      case Some(step) => {
+        logger.debug(s"...${step.getName} (${step.getId}): ${step.getStatus.getState}")
+        waitForJob(job)
+      }
+    }
+  }
+
+  /**
+   * Helper: runs a job and waits for the results to complete.
+   */
+  def runJobAndWait(steps: Seq[JobStep]): IO[Either[StepSummary, Unit]] = {
+    runJob(steps) >>= waitForJob
+  }
+
+  /**
+   * Helper: run a single step as a job.
+   */
+  def runStep(step: JobStep): IO[AddJobFlowStepsResult] = {
+    runJob(List(step))
+  }
+
+  /**
+   * Helper: run a single step and wait for it to complete.
+   */
+  def runStepAndWait(step: JobStep): IO[Either[StepSummary, Unit]] = {
+    runStep(step) >>= waitForJob
   }
 }
